+++ src/commands/open-actions-for-issue.ts
@@ -1,4 +1,3 @@
-import { RedmineServer } from "../redmine/redmine-server";
 import * as vscode from "vscode";
 import { IssueController } from "../controllers/issue-controller";
 import { ActionProperties } from "./action-properties";
@@ -8,15 +7,11 @@ export default async ({ server }: ActionProperties) => {
     placeHolder: `Type in issue id`,
   });
 
-  console.log(issueId);
-
   if (!issueId || !issueId.trim()) {
     return;
   }
 
-  console.log(issueId);
-
-  let promise = server.getIssueById(issueId);
+  const promise = server.getIssueById(issueId);
 
   console.log(issueId);
 
@@ -25,7 +20,7 @@ export default async ({ server }: ActionProperties) => {
       console.log(issue);
       if (!issue) return;
 
-      let controller = new IssueController(issue.issue, server);
+      const controller = new IssueController(issue.issue, server);
 
       controller.listActions();
     },