+++ src/redmine/models/time-entry.ts
@@ -1,0 +1,8 @@
+import { TimeEntryActivity } from "./time-entry-activity";
+
+export interface TimeEntry {
+  issue_id: number;
+  activity_id: TimeEntryActivity["id"];
+  hours: string;
+  comments: string;
+}